+++ auth-db/queries/email-auth.test.ts
@@ -12,41 +12,6 @@ describe("email-auth queries", () => {
     await db.users.deleteAll();
   });
 
-  describe("clearForgotPasswordToken", () => {
-    it("should clear forgot password token", async () => {
-      const user = await db.users.create({
-        email: "test@example.com",
-        createdAt: new Date(),
-      });
-
-      const passwordHash = Buffer.from([1, 2, 3]);
-      const token = "forgot-password-token";
-      const now = new Date();
-      now.setMilliseconds(0); // Round to seconds
-      const expiresAt = new Date(now.getTime() + 15 * 60 * 1000); // 15 minutes from now
-
-      await db.emailAuth.create({
-        userId: user.id,
-        email: user.email,
-        passwordHash,
-        forgotPasswordToken: token,
-        forgotPasswordTokenExpiresAt: expiresAt,
-      });
-
-      const updated = await db.emailAuth.clearForgotPasswordToken(user.id);
-      expect(updated).toMatchObject({
-        forgotPasswordToken: null,
-        forgotPasswordTokenExpiresAt: null,
-      });
-    });
-
-    it("should throw EmailAuthNotFoundError when user not found", async () => {
-      await expect(db.emailAuth.clearForgotPasswordToken(999)).rejects.toThrow(
-        EmailAuthNotFoundError,
-      );
-    });
-  });
-
   describe("verifyEmail", () => {
     it("should verify email and clear verification token", async () => {
       const user = await db.users.create({
