+++ auth-db/queries/email-auth/index.ts
@@ -5,3 +5,4 @@ export * from "./update-forgot-password-token.ts";
 export * from "./update-password-hash.ts";
 export * from "./get-by-forgot-password-token.ts";
 export * from "./update-password.ts";
+export * from "./clear-forgot-password-token.ts";
